organzie test suite into multiple folders
